fix(gate): split dispatch and classify out of the two god files grown by #1776
driver.rs (2705 > 2689) and registry.rs (2249 > 2214) went past their
ratcheted ceilings when #1776 merged over the db8cedc4 baseline
re-tighten. Move the one-dispatch-site scheduler
(Engine::execute_tool_calls) into driver/dispatch.rs and the file-op
classification pair (classify_file_ops / classify_file_op) into
registry/classify.rs — the settlement.rs / turn_probe.rs pattern —
instead of raising either ceiling. No behavior change.

Closes #1782
